I have refactored many things, including moving to a serviceLocator architecture for the main singletons (ui and raindropinstance), as I ran into problems with intialisation order of the UI.

+++ Assets/Plugins/LibreMetaverse/Catnip/Drawing/Color.cs
@@ -29,7 +29,7 @@ namespace Catnip.Drawing
         //
         // Summary:
         //     Red component of the color.
-        public float r;
+        public int r;
         //
         // Summary:
         //     Green component of the color.
